msm: smd_tty: Reset the open flag in close operation
The open flag in each SMD tty device is not reset during the close
operation. When the SMD tty device is closed and reopened quickly,
this leads to the completion of open operation even though the
underlying SMD channel is not open completely. Hence the few
initial operations on the SMD tty device fails. The open flag is
reset during the close operation, so that the successive open operation
waits until the underlying SMD channel is open completely.

Change-Id: I6c154a3c0fc9d084d4374c696fe05b8593999cf9
CRs-Fixed: 318810
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>