msm: smd_tty: Check the device is open before scheduling tasklets
When the SMD tty device is closed, tasklets to perform read operation
is killed first and then the underlying SMD channel is closed. There
is a race condition where the tasklet gets rescuduled between these
two operations. So ensure that the device is open before scheduling
tasklets to avoid this race condition.

Change-Id: I80c4e1538a69b5b05c1aadb0f4b0905f96a90001
CRs-Fixed: 318764
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>