msm: scm-pas: Request bandwidth for auth_and_reset()
The secure processor doesn't have the capability to request
bandwidth itself. Add bandwidth requests for it so that
performance doesn't suffer.

The bus driver is lacking a couple features:

 (1) No control of daytona fabric

 (2) No endpoints for the crypto engine on the daytona fabric

Solve (1) with a clock voter on daytona fabric and (2) with a
bandwidth request for SPS which lives on the daytona fabric.

Change-Id: Ib6ca67043bc3b97dbbecb1b571c4d33721fae11c
CRs-Fixed: 319454
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>