arm: irq: Correct set_irq_flags check for invalid irq
The current check compares against nr_irqs, which is assigned as
machine_desc->nr_irqs if defined. Per the usage in
arch_probe_nr_irqs(), nr_irqs is interpreted to mean the
preallocated irqs when used in a SPARSE_IRQ configuration. So
it's entirely possible that a chip driver will set its IRQ flags
for a IRQ number greater than nr_irqs.

NR_IRQS is still the largest IRQ any system can have, so let's
check for that instead.

Signed-off-by: Michael Bohan <mbohan@codeaurora.org>